release: v0.4.0

+++ CrumbDBCS/CrumbDBCS/Update.cs
@@ -4,7 +4,7 @@ namespace CrumbDBCS
 {
     public partial class CrumbDB
     {
-        public async Task<bool> Insert(string dirname, string documentname, string value, Encoding? encoding=null)
+        public async Task<bool> Update(string dirname, string documentname, string value, Encoding? encoding=null)
         {
             string filename = Path.Combine(dirname, $"{documentname}.json");
             SemaphoreSlim fileLock = GetFileLock(filename);
@@ -13,9 +13,18 @@ namespace CrumbDBCS
             try
             {
                 Directory.CreateDirectory(dirname);
-                Encoding fileEncoding = encoding ?? Encoding.UTF8;
-                await File.WriteAllTextAsync(filename, value, fileEncoding);
-                return true;
+
+                if (File.Exists(filename))
+                {
+
+                    Encoding fileEncoding = encoding ?? Encoding.UTF8;
+                    await File.WriteAllTextAsync(filename, value, fileEncoding);
+                    return true;
+                }
+                else
+                {
+                    return false;
+                }
             }
             catch (Exception)
             {
@@ -27,7 +36,7 @@ namespace CrumbDBCS
             }
         }
 
-        public async Task<bool> Insert(string dirname, string databasename, string collectionname, string documentname, string value, Encoding? encoding = null)
+        public async Task<bool> Update(string dirname, string databasename, string collectionname, string documentname, string value, Encoding? encoding = null)
         {
             string collectionDirname = Path.Combine(dirname, databasename, collectionname);
             string filename = Path.Combine(collectionDirname, $"{documentname}.json");
@@ -37,9 +46,18 @@ namespace CrumbDBCS
             try
             {
                 Directory.CreateDirectory(collectionDirname);
-                Encoding fileEncoding = encoding ?? Encoding.UTF8;
-                await File.WriteAllTextAsync(filename, value, fileEncoding);
-                return true;
+
+                if (File.Exists(filename))
+                {
+
+                    Encoding fileEncoding = encoding ?? Encoding.UTF8;
+                    await File.WriteAllTextAsync(filename, value, fileEncoding);
+                    return true;
+                }
+                else
+                {
+                    return false;
+                }
             }
             catch (Exception)
             {
